+++ apps/openmw/mwinput/inputmanagerimp.hpp
@@ -95,7 +95,7 @@ namespace MWInput
         std::string getControllerButtonIcon(int button) override;
         std::string getControllerAxisIcon(int axis) override;
 
-        int countSavedGameRecords() const override;
+        size_t countSavedGameRecords() const override;
         void write(ESM::ESMWriter& writer, Loading::Listener& progress) override;
         void readRecord(ESM::ESMReader& reader, uint32_t type) override;
